nhsbsa: Client for the NHS Business Services Authority Open Data Portal

A low-level client for the NHS Business Services Authority (NHSBSA) Open Data Portal https://opendata.nhsbsa.net, a 'CKAN' data catalogue. Provides thin wrappers around the portal's API actions for listing datasets, retrieving metadata, querying the datastore and downloading resource files. Results are returned as plain data (tibbles and lists) for the caller to interpret.

Further reading

The portal is a standard CKAN catalogue. For background on the API this package wraps, see:

The portal's API page: https://opendata.nhsbsa.net/pages/api

The CKAN Action API reference: https://docs.ckan.org/en/latest/api/

vignette("nhsbsa") explains how the package maps onto what you see on the portal website, and shows the different ways to query data.

See also

Author

Maintainer: Alasdair Warwick alasdair.warwick06@gmail.com (ORCID)